requirements.txt
PyTelegramBotAPI
python-dotenv
openpyxl
sqlite3